+++ aiotx/clients/__init__.py
@@ -4,6 +4,7 @@ import pkg_resources
 
 from ._evm_base_client import AioTxEVMClient
 from ._ton_base_client import AioTxTONClient
+from ._tron_base_client import AioTxTRONClient
 from ._utxo_base_client import AioTxUTXOClient
 
 __all__ = [
@@ -13,6 +14,7 @@ __all__ = [
     "AioTxPolygonClient",
     "AioTxBTCClient",
     "AioTxLTCClient",
+    "AioTxTRONClient",
 ]
 
 class AioTxBSCClient(AioTxEVMClient):