+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.7.0]
+- TRON client monitoring/wallet generation/get balance added
+- Node connection error is now RpcConnectionError
+- BTC init with failed get last block will do warning but not crash all code
+
 ## [2.6.0]
 - to_wei from_wei added hex numbers logic support / docs updated